Rename SpecUtil
